+++ modules/pytket-braket/docs/changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 ~~~~~~~~~
 
-0.17.0 (unreleased)
+0.17.0 (March 2022)
 -------------------
 
 * Add optional "region" argument for initializing BraketBackend.
@@ -9,6 +9,7 @@ Changelog
 * Add support for oqc provider.
 * Correct readout errors reported for Rigetti devices.
 * Add support for density-matrix simulators.
+* Update pytket version requirement to 1.0.
 
 0.16.0 (February 2022)
 ----------------------